Add a "prod" React build target.
This does not use Uglify or otherwise minify the output, but does run envify
in production mode.

This relies on the Closure Compiler to strip out the `"production" != "production"`
blocks, which it generally odes. The one place that it doesn't is with the
conditional inclusion of ReactTestUtils, so in that case we manually exclude
the addon altogether. While we're doing that also exclude other addons that
we're not using.

Overall this results in a 1.5K compiled JS increase (before Gzip, 552 bytes
after), so I think this is worth it given that stack traces will be much more
readable and we won't have to worry about manually quoting keys in the compiled
output.

CC=@milkbikis