Add a "prod" React build target.
This does not use Uglify or otherwise minify the output, but does run envify
in production mode.

This relies on the Closure Compiler to strip out the `"production" != "production"`
blocks, which it generally does. The one place that it doesn't is with the
conditional inclusion of ReactDebugTool, but exlcuding that is not as
easy as it was with 0.14 (see 38b78bc833ba5c2e8e84be3daeb36431b309d944)
for now it's left in.

Overall this results in a small compiled JS increase, so I think this is
worth it given that stack traces will be much more readable and we won't
have to worry about manually quoting keys in the compiled output.

Port of 61697fc8482b295ab3f417bb0c3d96cabd8cf816 to the 15-stable branch.
Does not include an "addons" version since we no longer use that output
file.